Update getting started instructions
Update the getting started instructions in the README.

I got Tarbell and this site up and running today and updated the
README to reflect things that could be more clear and linked to
portions of the Tarbell docs where appropriate.

I also removed the ``sudo`` from the ``pip`` command to install
Tarbell as you shouldn't need to use sudo when installing into a
virtualenv.