Basic d3 bar chart
Add a basic, reusable d3 chart function and use it to draw a
chart for convictions by category.

This could use some refinement in terms of colors and axis styles,
but it does provide a reasonable framework for drawing bar charts
and a namespace and reference implementation for drawing other
kinds of charts.

This implementation is based on the patterns described in these
articles:

* http://bost.ocks.org/mike/bar/3/
* http://bost.ocks.org/mike/chart/
* http://blog.apps.npr.org/2014/05/19/responsive-charts.html