Add about page
The main content is already pretty long.  Add a separate
"About" page as a place to put additional information, such as
credits, that is important, but doesn't fit within the core
narrative or content.

Move the credits file that @bepetersn created from the ``slides``
subdirectory and into the root of the project.

Include the credits file in ``about.html`` and remove the content
from the intro slide.

In order to make the hash-based URLs on the main index page,
introduce the ``nav_prefix`` template variable.  This needs to be
set to 'index.html' in any page templates that aren't 'index.html'.