Add footer
Add a fixed footer with credit information at the bottom of the page.

I don't think we need to implement the expanding behavior as it
adds complexity without a lot of value.  The core information fits
nicely in the thin footer bar.  Other info can go on the about
page.

Fixes #75