Add a border around the city of Chicago
Make the distinction between the Chicago and the rest of the map
with a bold black line.

This line is represented as a LineString geometry in the GeoJSON
file ``data/chicago.json``.

Rename some of the models in ``convictions.js`` to generalize cases
since the map view is now loading two GeoJSON files.

Update ``CommunityAreaMapView`` to draw both the community area
boundary and the City of Chicago boundary.

Addresses https://github.com/sc3/cook-convictions/issues/66