Suburbs and community areas on one map
It was silly to make these two separate maps.  Just add the
suburban places as a layer and make a single map.

Rename some of the map view classes, functions and variables to
reflect this change.

Remove the slide for the suburbs map.

Addresses https://github.com/sc3/cook-convictions/issues/77