Stacked bar chart
Implement stacked bar chart for classes of drug convictions.

Define JSON for data in index.html.

Refactor some of the logic in barChart to allow overriding it
in the stacked bar chart.  The big reason for this is that the
rectangles are now nested under a ``g`` element instead of being
at the top level.

Set fill color of bar charts in code instead of CSS.

This uses a qualitative set of colors from Color Brewer.  I tried
using the same scale as the map, but the gradient-like values of
the same black/gray hues didn't allow us to easily distinguish
between the bar stacks.

Addresses #60.