Lower initial zoom level of map
Use an initial zoom level of 10 instead of 11 for the map so that
the entire city is visible in the viewport and so that it's clear
that the map also includes suburban areas.

Also add JavaScript to dynamically size the map container to the
full height of the viewport and to resize it when the window is
resized.

Fixes https://github.com/sc3/cook-convictions/issues/85