Fix the heading for the convictions by category chart
We didn't categorize every conviction, so I'm not sure if it's
correct to say that drug convictions account for the largest
"chunk" of all convictions in our data.  Change the heading so
it describes the chart below, only describing drug convictions
relative to our other categories.

Fixes https://github.com/sc3/cook-convictions/issues/108