Remove dispositions stat
Remove dispositions stat from the area with the bigger icons.  Move
it to the "Windy City Convictions by the Numbers" section.

Move headings above the icons and make the font somewhat larger.

Add commas to stats. A custom Jinja filter was needed for this and I
added it in tarbell_config.py using the pattern at the bottom of the
"Configuring Tarbell" section of http://tarbell.tribapps.com/readme/.

Fixes https://github.com/sc3/cook-convictions/issues/119