Make slide titles, filenames, ids line up with sheet
Make the slide titles, filenames and ids line up with the spreadsheet
values.  To do this, remove the headings from the template files and
instead render them in index.html.  Add a ``show_title`` column to
the Tarbell spreadsheet to hide the title display for a few slides.
Update the spreadsheet to use the heading titles that were in the
template files.  Rename some of the slide ids so they more closely
follow the titles.  Rename the template files so they match the new
slide names.

Remove quick edit links from slides in preparation for launch.

Fixes https://github.com/sc3/cook-convictions/issues/104